Feat: current location hook
